refactor(i18n): replace CliConfig with EditLanguageConfig for language management
